Add UX improvements to attachments
