+++ example/src/comments-list/index.js
@@ -1,34 +1,68 @@
 import React from "react";
+import zacs from "@nozbe/zacs";
 
 import Box from "../shared/box";
 import ListHeader from "../shared/list-header";
 import List from "../shared/list";
 import InputWithButton from "../shared/input-with-button";
 import { useState } from "react/cjs/react.development";
-import AddAttachment from "./add-attachment";
-import Button from "../shared/button/button";
+
+import style from "./style.module.css";
+
+const Label = zacs.styled("label", style.label);
+const Input = zacs.styled("input", style.input);
 
 const CommentsList = ({ comments, addComment, addAttachmentsByFilesArray }) => {
-  const [attachmentsFormVisible, setAttachmentsFormVisible] = useState(false);
+  const [files, setFiles] = useState([]);
+  const [comment, setComment] = useState("");
+
+  const handleAddFile = (event) => {
+    const { files: formFiles } = event.target;
+    if (formFiles) {
+      setFiles([...formFiles]);
+    }
+  };
+
+  const handleAddComment = () => {
+    if (files.length) {
+      addAttachmentsByFilesArray(comment, files);
+      setComment("");
+      setFiles([]);
+      return;
+    }
+
+    addComment(comment);
+    setComment("");
+  };
+
+  const filesArray = Array.from(files);
 
   return comments ? (
     <Box>
       <ListHeader>Comments:</ListHeader>
       <List items={comments} itemContentKey="body" />
-      {!attachmentsFormVisible && (
-        <>
-          <InputWithButton onClick={addComment} btnCaption="Add comment" />
-          <Button
-            type="submit"
-            onClick={() => setAttachmentsFormVisible(!attachmentsFormVisible)}
-          >
-            Show attachments form
-          </Button>
-        </>
-      )}
-      {attachmentsFormVisible && (
-        <AddAttachment {...{ addAttachmentsByFilesArray }} />
-      )}
+
+      <InputWithButton
+        onClick={handleAddComment}
+        value={comment}
+        onChange={setComment}
+        btnCaption="Add comment"
+      />
+
+      <Input
+        type="file"
+        name="fileInput"
+        id="fileInput"
+        multiple
+        onChange={handleAddFile}
+      />
+
+      <Label htmlFor="fileInput">Select files</Label>
+      <ul>
+        {filesArray.map((file) => (
+          <li>{file.name}</li>
+        ))}
+      </ul>
     </Box>
   ) : null;
 };
