+++ example/src/comments-list/style.module.css
@@ -15,4 +15,5 @@
 
 .input {
   visibility: hidden;
+  position: absolute;
 }
